return `false` from isExec on any os.Stat error
os.Stat could return a set of possible instances of *os.PathError, but not necessary always os.ErrExist or os.ErrNotExist, wich are what os.IsNotExist tests for.
This is observed when passing, for example, "test/foo" as argument, for which os.Stat will return error whose string print is "not a directory" because the executable "test" is present in /bin/test but it is not a directory where "foo" can be found under. Thus the error evades the check os.IsNotExist check, falls down to the fi.IsDir(), where panics with nil dereference error due to "fi" being nil because os.Stat returned an error.